+++ linphone-app/ui/views/App/Calls/CallsWindow.js
@@ -82,7 +82,7 @@ function getContent (call, conferenceInfoModel) {
 	}
 	var CallModel = Linphone.CallModel
 	if (status === CallModel.CallStatusIncoming) {
-		return null;
+		return incall;
 	}
 	if( window.conferenceInfoModel != call.conferenceInfoModel) {
 		Qt.callLater(function(){window.conferenceInfoModel = call.conferenceInfoModel})