+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 5.0.10 - undefined
 
 ### Fixed
-- Blank notification when declining incoming call.
+- Remove blank notification when declining incoming call.
+- Remove blank page when opening calls window and add a waiting step while connecting to call.
+- Camera activation issue based on call status.
 
 ## 5.0.9 - 2023-01-31